pyproject.toml
[project]
name = "missionorch"
version = "1.0.0"
description = "A military course of action (COA) generation and validation system"
authors = [{name = "Developer", email = "developer@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=1.8.0",
    "pyyaml>=5.4.0",
    "openai>=1.0.0",
    "google-generativeai>=0.1.0",
    "anthropic>=0.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0"
]

[tool.setuptools.packages.find]
where = ["src"]